refactor: move the Inferenceset controller to a top-level pkg
